Refactor the way the executor processes steps
