+++ src/utils/constants.js
@@ -1,5 +1,6 @@
 import path from 'path'
 
+
 const TX_LOOKUP_TIME = 5000 // milliseconds
 const HD_WALLET_ACCOUNTS = 10
 const ETH_ADDRESS_LENGTH = 42
@@ -7,7 +8,7 @@ const DEFAULT_CONFIG_FILE_PATH = path.join(__dirname, '../../conf/config.json')
 const DEFAULT_MARKET_FILE_PATH = path.join(__dirname, '../../conf/markets.json')
 const DEFAULT_GAS_PRICE = 4000000000 // 4 GWEI
 const DEFAULT_GAS_LIMIT = 7500000
-const SDK_VERSION = '1.4.0'
+const SDK_VERSION = require('../../package.json').version
 
 const MARKET_STAGES = {
   created: 0,
@@ -15,14 +16,21 @@ const MARKET_STAGES = {
   closed: 2
 }
 
+const EXECUTION_TYPES = {
+  deploy: 'deploy',
+  resolve: 'resolve'
+}
+
+
 module.exports = {
-  TX_LOOKUP_TIME,
-  ETH_ADDRESS_LENGTH,
   DEFAULT_CONFIG_FILE_PATH,
   DEFAULT_MARKET_FILE_PATH,
   DEFAULT_GAS_PRICE,
   DEFAULT_GAS_LIMIT,
-  MARKET_STAGES,
+  ETH_ADDRESS_LENGTH,
+  EXECUTION_TYPES,
   HD_WALLET_ACCOUNTS,
-  SDK_VERSION
+  MARKET_STAGES,
+  SDK_VERSION,
+  TX_LOOKUP_TIME
 }
